# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  serialize.InfobaseInfo:
    properties:
      cluster_id:
        example: efa3672f-947a-4d84-bd58-b21997b83561
        type: string
      date_offset:
        example: 0
        type: integer
      db_name:
        example: base
        type: string
      db_pwd:
        example: password
        type: string
      db_server:
        example: sql
        type: string
      db_user:
        example: user
        type: string
      dbms:
        example: MSSQLServer
        type: string
      denied_from:
        example: "2020-10-01T08:30:00Z"
        type: string
      denied_message:
        example: Выполняется обновление базы
        type: string
      denied_parameter:
        example: "123"
        type: string
      denied_to:
        example: "2020-10-01T08:30:00Z"
        type: string
      descr:
        example: Это очень хорошая база
        type: string
      external_session_manager_connection_string:
        example: http://auth2.com
        type: string
      external_session_manager_required:
        example: false
        type: boolean
      license_distribution:
        example: 0
        type: integer
      locale:
        example: ru_RU
        type: string
      name:
        example: УППБоеваяБаза
        type: string
      permission_code:
        example: "123"
        type: string
      reserve_working_processes:
        example: false
        type: boolean
      safe_mode_security_profile_name:
        example: profile1
        type: string
      scheduled_jobs_deny:
        example: false
        type: boolean
      security_level:
        example: 0
        type: integer
      security_profile_name:
        example: sec_profile1
        type: string
      sessions_deny:
        example: false
        type: boolean
      uuid:
        example: efa3672f-947a-4d84-bd58-b21997b83561
        type: string
    type: object
  serialize.InfobaseSummaryInfo:
    properties:
      cluster_id:
        type: string
      descr:
        example: УППБоеваяБаза
        type: string
      name:
        example: УППБоеваяБаза
        type: string
      uuid:
        example: efa3672f-947a-4d84-bd58-b21997b83561
        type: string
    type: object
...
host: localhost:3001
info:
  contact:
    email: khorevaa@yandex.ru
    name: API Support
  description: KUBOdin Swagger UI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'KUBOdin: Remote Administration for 1S.Enterprise Application Servers'
  version: "1.0"
paths:
...
  /app/{app}/clusters/{cluster}/infobases:
    get:
      consumes:
      - application/json
      description: получение списка информационных баз с кластера
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serialize.InfobaseSummaryInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: получение списка информационных баз с кластера
      tags:
      - infobases
    post:
      consumes:
      - application/json
      description: оздание информационной базы на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: create server db
        in: query
        name: create-db
        type: boolean
      - description: new info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/serialize.InfobaseInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.InfobaseInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: оздание информационной базы на кластере
      tags:
      - infobases
  /app/{app}/clusters/{cluster}/infobases/{infobase}:
    delete:
      consumes:
      - application/json
      description: удаляет информационную базу с сервера кластера
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase name or uuid
        in: path
        name: infobase
        required: true
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: delete server db
        in: query
        name: delete-db
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: удаляет информационную базу с сервера кластера
      tags:
      - infobases
    get:
      consumes:
      - application/json
      description: получение информации об информационной базе с кластера
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase name or uuid
        in: path
        name: infobase
        required: true
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.InfobaseInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: получение информации об информационной базе с кластера
      tags:
      - infobases
    post:
      consumes:
      - application/json
      description: побновление информации об информационной базе на кластере
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: path
        name: cluster
        required: true
        type: string
      - description: infobase name or uuid
        in: path
        name: infobase
        required: true
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      - description: new info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/serialize.InfobaseInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: обновление информации об информационной базе на кластере
      tags:
      - infobases
...
  /app/{app}/infobases:
    get:
      consumes:
      - application/json
      description: получение списка информационных баз с сервера 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serialize.InfobaseSummaryInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Response'
      summary: получение списка информационных баз с сервера 1С Предприятие
      tags:
      - infobases
    post:
      consumes:
      - application/json
      description: создание информационной базы на сервере 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: create server db
        in: query
        name: create-db
        type: boolean
      - description: new info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/serialize.InfobaseInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.InfobaseInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: создание информационной базы на сервере 1С Предприятие
      tags:
      - infobases
  /app/{app}/infobases/{infobase}:
    delete:
      consumes:
      - application/json
      description: удаляет информационную базу с сервера 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: infobase name or uuid
        in: path
        name: infobase
        required: true
        type: string
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: delete server db
        in: query
        name: delete-db
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: удаляет информационную базу с сервера 1С Предприятие
      tags:
      - infobases
    get:
      consumes:
      - application/json
      description: получение информации об информационной базе с сервера 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: infobase name or uuid
        in: path
        name: infobase
        required: true
        type: string
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  $ref: '#/definitions/serialize.InfobaseInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: получение информации об информационной базе с сервера 1С Предприятие
      tags:
      - infobases
    post:
      consumes:
      - application/json
      description: обновление информации об информационной базе на сервера 1С Предприятие
      parameters:
      - description: app name
        in: path
        name: app
        required: true
        type: string
      - description: infobase name or uuid
        in: path
        name: infobase
        required: true
        type: string
      - description: cluster uuid
        in: query
        name: cluster-id
        type: string
      - description: cluster user
        in: query
        name: cluster-usr
        type: string
      - description: cluster password
        in: query
        name: cluster-pwd
        type: string
      - description: infobase user
        in: query
        name: infobase-usr
        type: string
      - description: infobase password
        in: query
        name: infobase-pwd
        type: string
      - description: force update ignore cache
        in: query
        name: force
        type: boolean
      - description: new info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/serialize.InfobaseInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/Response'
            - properties:
                data:
                  type: string
              type: object
      summary: обновление информации об информационной базе на сервера 1С Предприятие
      tags:
      - infobases
...
swagger: "2.0"